Kernel: Xiaomi kernel changes for Xiaomi Civi Android R
The Patch based on QualComm release TAG:LA.UM.9.14.r1-11500-LAHAINA.0
The kernel config file used is mona_QGKI.config

+++ drivers/dax/bus.h
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0
 /* Copyright(c) 2016 - 2018 Intel Corporation. All rights reserved. */
+/* Copyright (C) 2021 XiaoMi, Inc. */
 #ifndef __DAX_BUS_H__
 #define __DAX_BUS_H__
 #include <linux/device.h>
@@ -11,7 +12,7 @@ struct dax_region;
 void dax_region_put(struct dax_region *dax_region);
 struct dax_region *alloc_dax_region(struct device *parent, int region_id,
 		struct resource *res, int target_node, unsigned int align,
-		unsigned long flags);
+		unsigned long long flags);
 
 enum dev_dax_subsys {
 	DEV_DAX_BUS,
