+++ net/openvswitch/meter.h
@@ -1,6 +1,7 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
 /*
  * Copyright (c) 2017 Nicira, Inc.
+ * Copyright (C) 2021 XiaoMi, Inc.
  */
 
 #ifndef METER_H
@@ -23,7 +24,7 @@ struct dp_meter_band {
 	u32 type;
 	u32 rate;
 	u32 burst_size;
-	u32 bucket; /* 1/1000 packets, or in bits */
+	u64 bucket; /* 1/1000 packets, or in bits */
 	struct ovs_flow_stats stats;
 };
